// apps/web-stride/src/graphql/types/general_types.tsx
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
const defaultOptions: {[key: string]: any} = {};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  ActionCoin: any;
  ActionDelegation: any;
  ActionPagination: any;
  ActionRedelegation: any;
  ActionUnbondingDelegation: any;
  _coin: any;
  _dec_coin: any;
  _text: any;
  bigint: any;
  jsonb: any;
  numeric: any;
  smallint: any;
  timestamp: any;
};
// ...
export type Distribution_Params = {
  __typename?: 'distribution_params';
  height: Scalars['bigint'];
  one_row_id: Scalars['Boolean'];
  params: Scalars['jsonb'];
};
// ...
export type Gov_Params = {
  __typename?: 'gov_params';
  deposit_params: Scalars['jsonb'];
  height: Scalars['bigint'];
  one_row_id: Scalars['Boolean'];
  tally_params: Scalars['jsonb'];
  voting_params: Scalars['jsonb'];
};

/** columns and relationships of "gov_params" */
export type Stakeibc_Params = {
  __typename?: 'stakeibc_params';
  buffer_size: Maybe<Scalars['bigint']>;
  deposit_interval: Maybe<Scalars['bigint']>;
  rewards_interval: Maybe<Scalars['bigint']>;
  delegate_interval: Maybe<Scalars['bigint']>;
  ica_timeout_nanos: Maybe<Scalars['bigint']>;
  reinvest_interval: Maybe<Scalars['bigint']>;
  stride_commission: Maybe<Scalars['bigint']>;
  ibc_timeout_blocks: Maybe<Scalars['bigint']>;
  redemption_rate_interval: Maybe<Scalars['bigint']>;
  fee_transfer_timeout_nanos: Maybe<Scalars['bigint']>;
  ibc_transfer_timeout_nanos: Maybe<Scalars['bigint']>;
  max_stake_ica_calls_per_epoch: Maybe<Scalars['bigint']>;
  validator_rebalancing_threshold: Maybe<Scalars['bigint']>;
  safety_max_redemption_rate_threshold: Maybe<Scalars['bigint']>;
  safety_min_redemption_rate_threshold: Maybe<Scalars['bigint']>;
};
// ...
export type Mint_Params = {
  __typename?: 'mint_params';
  height: Scalars['bigint'];
  one_row_id: Scalars['Boolean'];
  params: Scalars['jsonb'];
};
// ...
export type Slashing_Params = {
  __typename?: 'slashing_params';
  height: Scalars['bigint'];
  one_row_id: Scalars['Boolean'];
  params: Scalars['jsonb'];
};
// ...
export type Staking_Params = {
  __typename?: 'staking_params';
  height: Scalars['bigint'];
  one_row_id: Scalars['Boolean'];
  params: Scalars['jsonb'];
};
// ...
export type ParamsQuery = {
  stakingParams: Array<{ __typename?: 'staking_params' } & Pick<Staking_Params, 'params'>>;
  slashingParams: Array<{ __typename?: 'slashing_params' } & Pick<Slashing_Params, 'params'>>;
  mintParams: Array<{ __typename?: 'mint_params' } & Pick<Mint_Params, 'params'>>;
  distributionParams: Array<
    { __typename?: 'distribution_params' } & Pick<Distribution_Params, 'params'>
  >;
  govParams: Array<
    { __typename?: 'gov_params' } & {
      depositParams: Gov_Params['deposit_params'];
      tallyParams: Gov_Params['tally_params'];
      votingParams: Gov_Params['voting_params'];
    }
  >;
  stakeibcParams: Array<
    { __typename?: 'stakeibc_params' } & {
      params: {
        buffer_size: Stakeibc_Params['buffer_size'];
        deposit_interval: Stakeibc_Params['deposit_interval'];
        rewards_interval: Stakeibc_Params['rewards_interval'];
        delegate_interval: Stakeibc_Params['delegate_interval'];
        ica_timeout_nanos: Stakeibc_Params['ica_timeout_nanos'];
        reinvest_interval: Stakeibc_Params['reinvest_interval'];
        stride_commission: Stakeibc_Params['stride_commission'];
        ibc_timeout_blocks: Stakeibc_Params['ibc_timeout_blocks'];
        redemption_rate_interval: Stakeibc_Params['redemption_rate_interval'];
        fee_transfer_timeout_nanos: Stakeibc_Params['fee_transfer_timeout_nanos'];
        ibc_transfer_timeout_nanos: Stakeibc_Params['ibc_transfer_timeout_nanos'];
        max_stake_ica_calls_per_epoch: Stakeibc_Params['max_stake_ica_calls_per_epoch'];
        validator_rebalancing_threshold: Stakeibc_Params['validator_rebalancing_threshold'];
        safety_max_redemption_rate_threshold: Stakeibc_Params['safety_max_redemption_rate_threshold'];
        safety_min_redemption_rate_threshold: Stakeibc_Params['safety_min_redemption_rate_threshold'];
      };
    }
  >;
};